fix badge count logic
