+++ Example/BasicExample/TwilioEngageExampleNotificationServiceExtension/Info.plist
@@ -4,13 +4,6 @@
 <dict>
 	<key>NSExtension</key>
 	<dict>
-		<key>NSExtensionAttributes</key>
-		<dict>
-			<key>UNNotificationExtensionCategory</key>
-			<string>open_url</string>
-			<key>UNNotificationExtensionUserInteractionEnabled</key>
-			<true/>
-		</dict>
 		<key>NSExtensionPointIdentifier</key>
 		<string>com.apple.usernotifications.service</string>
 		<key>NSExtensionPrincipalClass</key>
